Fixed a problem in the manager where combine_lflat/dark/bias is called, but the corresponding files don't exist.
Previously, if `combine_bias/dark/lflat` was called when there were no corresponding files (regardless of whether `reduce_bias/dark/lflat` had been called), the package would make a link to a non-existant file. Subsequent invocations would find this file, but os.path.exists returns False (broken link), but os.symlink errors because there is already a symlink in place.

This corrects the logic such that (a) the broken symlink will not be created, and (b) if the symlink is there but broken and `overwrite=True`, then the existing link will be removed.

+++ manager.py
@@ -1217,9 +1217,10 @@ class Manager:
                 for link_dir in self.reduced_dirs(dir_type, ccd=ccd,
                                                   do_not_use=False):
                     link_path = os.path.join(link_dir, filename)
-                    if overwrite and os.path.exists(link_path):
+                    if overwrite and (os.path.exists(link_path) or
+                                      os.path.islink(link_path)):
                         os.remove(link_path)
-                    if not os.path.exists(link_path):
+                    if (not os.path.exists(link_path)) and os.path.exists(path):
                         os.symlink(os.path.relpath(path, link_dir),
                                    link_path)
         return
